--- gravitee-node-cluster/gravitee-node-cluster-plugin-hazelcast-provider/pom.xml
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright © 2015 The Gravitee team (http://gravitee.io)
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>io.gravitee.node</groupId>
-        <artifactId>gravitee-node-cluster</artifactId>
-        <version>9.0.0-alpha.5</version>
-    </parent>
-
-    <artifactId>gravitee-node-cluster-plugin-hazelcast-provider</artifactId>
-    <name>Gravitee.io - Node - Cluster - Plugin - Hazelcast Provider</name>
-    <description>Owns the embedded Hazelcast instance and exposes it via DistributedMapProvider. Activated independently of cluster.type so operators can use Hazelcast-backed distributed maps without also running APIM cluster sync on Hazelcast.</description>
-
-    <properties>
-        <!-- Property used by the publication job in CI-->
-        <publish-folder-path>plugins/node-cluster</publish-folder-path>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>io.gravitee.node</groupId>
-            <artifactId>gravitee-node-api</artifactId>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>io.gravitee.node</groupId>
-            <artifactId>gravitee-node-logging</artifactId>
-            <scope>provided</scope>
-        </dependency>
-        <!-- Hazelcast -->
-        <dependency>
-            <groupId>com.hazelcast</groupId>
-            <artifactId>hazelcast</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>org.awaitility</groupId>
-            <artifactId>awaitility</artifactId>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <resources>
-            <resource>
-                <directory>src/main/resources</directory>
-                <filtering>true</filtering>
-            </resource>
-        </resources>
-
-        <plugins>
-            <plugin>
-                <artifactId>maven-assembly-plugin</artifactId>
-            </plugin>
-        </plugins>
-    </build>
-</project>
