--- gravitee-node-cluster/gravitee-node-cluster-plugin-hazelcast/src/main/assembly/plugin-assembly.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" ?>
-<!--
-
-    Copyright © 2015 The Gravitee team (http://gravitee.io)
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<assembly>
-	<id>plugin</id>
-	<formats>
-		<format>zip</format>
-	</formats>
-	<includeBaseDirectory>false</includeBaseDirectory>
-
-	<!-- Include the main plugin Jar file -->
-	<files>
-		<file>
-			<source>${project.build.directory}/${project.build.finalName}.jar</source>
-		</file>
-	</files>
-
-	<!-- Include plugin schema configuration -->
-	<fileSets>
-		<fileSet>
-			<directory>src/main/resources/schemas</directory>
-			<outputDirectory>schemas</outputDirectory>
-		</fileSet>
-	</fileSets>
-
-	<!-- Finally include plugin dependencies -->
-	<dependencySets>
-		<dependencySet>
-			<outputDirectory>lib</outputDirectory>
-			<useProjectArtifact>false</useProjectArtifact>
-		</dependencySet>
-	</dependencySets>
-</assembly>
