--- gravitee-node-cluster/gravitee-node-cluster-plugin-hazelcast/src/main/resources/plugin.properties
@@ -1,6 +1,0 @@
-id=cluster-hazelcast
-name=Hazelcast Cluster
-version=${project.version}
-description=${project.description}
-class=io.gravitee.node.plugin.cluster.hazelcast.HazelcastClusterManager
-type=cluster
