--- gravitee-node-cluster/gravitee-node-cluster-plugin-hazelcast/src/test/resources/cluster.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-  ~ /**
-  ~  * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
-  ~  *
-  ~  * Licensed under the Apache License, Version 2.0 (the "License");
-  ~  * you may not use this file except in compliance with the License.
-  ~  * You may obtain a copy of the License at
-  ~  *
-  ~  *         http://www.apache.org/licenses/LICENSE-2.0
-  ~  *
-  ~  * Unless required by applicable law or agreed to in writing, software
-  ~  * distributed under the License is distributed on an "AS IS" BASIS,
-  ~  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~  * See the License for the specific language governing permissions and
-  ~  * limitations under the License.
-  ~  */
-  -->
-
-<!--
-    The following is given as a simple example to enable hazelcast cluster on local environment.
-    It will be used when cluster.type: hazelcast is set in gravitee.yml.
-
-    For advanced configuration, please refer to official documentation: https://docs.hazelcast.com/imdg/4.2/clusters/discovery-mechanisms
--->
-<hazelcast xmlns="http://www.hazelcast.com/schema/config"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="http://www.hazelcast.com/schema/config
-           http://www.hazelcast.com/schema/config/hazelcast-config-5.3.xsd">
-
-    <cluster-name>gio-apim-gateway</cluster-name>
-    <network>
-        <port auto-increment="true" port-count="100">5701</port>
-        <join>
-            <auto-detection enabled="true"/>
-            <multicast enabled="false"/>
-            <tcp-ip enabled="true">
-                <interface>127.0.0.1</interface>
-            </tcp-ip>
-        </join>
-    </network>
-</hazelcast>
