+++ gravitee-node-plugin-hazelcast/src/main/assembly/plugin-assembly.xml
