+++ gravitee-node-plugin-hazelcast/src/main/java/io/gravitee/node/plugin/cluster/hazelcast/provider/HazelcastDistributedMapProvider.java
@@ -1,0 +1,40 @@
+/**
+ * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.gravitee.node.plugin.cluster.hazelcast.provider;
+
+import com.hazelcast.core.HazelcastInstance;
+import com.hazelcast.map.IMap;
+import io.gravitee.node.api.cluster.DistributedMap;
+import io.gravitee.node.api.cluster.DistributedMapProvider;
+import lombok.RequiredArgsConstructor;
+
+/**
+ * Holds a {@link HazelcastInstance} through Spring's {@code @Lazy} proxy so construction of this
+ * provider does not boot Hazelcast. The proxy dereferences on the first {@link #get(String)} call,
+ * so when no consumer ever asks for a {@link DistributedMap} (e.g. {@code cluster.type=standalone}
+ * + {@code ratelimit.type!=hazelcast}), Hazelcast stays dormant.
+ */
+@RequiredArgsConstructor
+public class HazelcastDistributedMapProvider implements DistributedMapProvider {
+
+    private final HazelcastInstance hazelcastInstance;
+
+    @Override
+    public <K, V> DistributedMap<K, V> get(final String name) {
+        IMap<K, V> iMap = hazelcastInstance.getMap(name);
+        return new HazelcastDistributedMap<>(iMap);
+    }
+}
