+++ gravitee-node-plugin-hazelcast/src/main/resources/plugin.properties
@@ -1,0 +1,9 @@
+id=hazelcast-provider
+name=Hazelcast Provider
+version=${project.version}
+description=${project.description}
+class=io.gravitee.node.plugin.cluster.hazelcast.provider.HazelcastProvider
+type=hazelcast-provider
+# Priority below 1000 so this plugin deploys before cluster/cache/repository plugins
+# that may depend on its HazelcastInstance + DistributedMapProvider beans.
+priority=500
