+++ gravitee-node-plugin-hazelcast/src/test/resources/cluster.xml
