+++ pom.xml
@@ -39,6 +39,7 @@
         <module>gravitee-node-api</module>
         <module>gravitee-node-cache</module>
         <module>gravitee-node-cluster</module>
+        <module>gravitee-node-plugin-hazelcast</module>
         <module>gravitee-node-opentelemetry</module>
         <module>gravitee-node-vertx</module>
         <module>gravitee-node-management</module>
